remove several older geocoding functions, use lut instead
